requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.7
gunicorn==21.2.0
pytest-django==4.7.0
sqlparse==0.4.4
typing_extensions==4.8.0
async-timeout==4.0.3 
channels_redis==4.1.0 
msgpack==1.0.7 
redis==5.0.1